address review comments
thanks Andy!
