+++ rust/opendp/src/lib.rs
@@ -140,6 +140,7 @@
 #![allow(clippy::upper_case_acronyms)]
 #![allow(clippy::just_underscores_and_digits)]
 #![allow(clippy::type_complexity)]
+#![recursion_limit="512"]
 
 // create clones of variables that are free to be consumed by a closure
 macro_rules! enclose {
